[Enhance] Add support for Linux

+++ uninstall.sh
@@ -1,6 +1,10 @@
-#!/bin/sh
+#!/bin/bash
 #
 # Uninstall.
+
+DIR_PATH=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
+
+source $DIR_PATH/sh_utils/utils.sh
 
 main() {
   uninstall_tools
@@ -11,8 +15,10 @@ main() {
 }
 
 uninstall_tools() {
-  brew remove kubectx hub shfmt
-  brew remove tmux wget ffmpeg
+  if check_os $OS_MAC; then
+    brew remove kubectx hub shfmt
+    brew remove tmux wget ffmpeg
+  fi
 }
 
 uninstall_shell() {
@@ -20,26 +26,24 @@ uninstall_shell() {
 }
 
 uninstall_languages() {
-  brew remove go
-  brew remove node yarn
+  if check_os $OS_MAC; then
+    brew remove go
+    brew remove node yarn
 
-  brew remove --cask adoptopenjdk
+    brew remove --cask adoptopenjdk
+  fi
 
   rustup self uninstall
 }
 
 uninstall_homebrew() {
+  if ! check_os $OS_MAC; then
+    return
+  fi
+
   if check_cmd brew; then
     /usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/uninstall)"
   fi
-}
-
-#
-# Utils
-#
-
-check_cmd() {
-  command -v "$1" >/dev/null 2>&1
 }
 
 main "$@"
